+++ src/lib.rs
@@ -220,3 +220,9 @@ pub use validator::{DefaultValidator, ValidationResult, Validator};
 
 mod menu;
 pub use menu::{CompletionMenu, HistoryMenu, Menu};
+
+mod internal;
+pub use internal::{
+    get_all_keybinding_info, get_reedline_edit_commands, get_reedline_keybinding_modifiers,
+    get_reedline_keycodes, get_reedline_prompt_edit_modes, get_reedline_reedline_events,
+};
